+++ templates/navbar.html
@@ -7,10 +7,10 @@
     <div class="collapse navbar-collapse" id="navbarSupportedContent">
       <ul class="navbar-nav me-auto mb-2 mb-lg-0">
         <li class="nav-item">
-          <a class="nav-link active" aria-current="page" href="{{ url_for('user_profile', name='masa')}}">Profile</a>
+          <a class="nav-link active" aria-current="page" href="{{ url_for('user.user_profile', name='masa')}}">Profile</a>
         </li>
         <li class="nav-item">
-          <a class="nav-link active" aria-current="page" href="{{ url_for('add_user')}}">Add User</a>
+          <a class="nav-link active" aria-current="page" href="{{ url_for('user.add_user')}}">Add User</a>
         </li>
 
       </ul>
